move the speculative cp code out of the shared distributed layer
