+++ nemo_automodel/components/speculative/dspark/target.py
@@ -65,10 +65,8 @@ class HFDSparkTargetModel:
         self.cp_mesh = cp_mesh
         self._cp_size = cp_mesh.size() if cp_mesh is not None else 1
         if self._cp_size > 1:
-            from nemo_automodel.components.distributed.cp_utils import (
-                attach_context_parallel_hooks,
-                attach_cp_kv_gather_hooks,
-            )
+            from nemo_automodel.components.distributed.cp_utils import attach_context_parallel_hooks
+            from nemo_automodel.components.speculative.target_cp import attach_cp_kv_gather_hooks
 
             # Strip the 4D mask (self_attn then calls SDPA on the local shard), and
             # all-gather K/V so each rank attends its local Q against the full sequence
@@ -188,7 +186,7 @@ class HFDSparkTargetModel:
         if self._cp_size > 1:
             # Shard the sequence, run the target as ring attention, then gather every
             # captured hidden state back to the full sequence.
-            from nemo_automodel.components.distributed.cp_utils import run_target_cp_forward_and_gather
+            from nemo_automodel.components.speculative.target_cp import run_target_cp_forward_and_gather
 
             keys: list = []
