+++ nemo_automodel/components/speculative/eagle/zigzag_ring_attention.py
